pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eh-system"
version = "0.1.0"
description = "Elektronik Harp Arayüz Sistemi — SDR tabanlı EH masaüstü uygulaması"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "PySide6>=6.6",
    "pyqtgraph>=0.13",
    "numpy>=1.24",
    "pyzmq>=25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-qt>=4.2",
]

[project.scripts]
eh-system = "ehapp.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
